backend/requirements.txt
# backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
websockets==12.0
python-multipart==0.0.6
openai==1.12.0
python-dotenv==1.0.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
pydantic  # Downgraded for FastAPI compatibility
docling-core  # Replace with the last compatible version
mistralai>=0.0.12  # Versão específica que contém a API atualizada
requests>=2.28
langchain-docling==0.2.0
langchain-core
regex>=2023.10.3